standalone-personal-accounting/backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6
pytesseract>=0.3.10
Pillow>=10.0.0
reportlab>=4.0.0
